move FavaModule back to fava.core
