+++ fava/core/file.py
@@ -22,8 +22,8 @@ from beancount.parser.printer import format_entry  # type: ignore
 from fava.core.fava_options import InsertEntryOption
 from fava.core.filters import get_entry_accounts
 from fava.core.misc import align
+from fava.core.module_base import FavaModule
 from fava.helpers import FavaAPIException
-from fava.helpers import FavaModule
 
 
 #: The flags to exclude when rendering entries entries.
