+++ fava/helpers.py
@@ -20,13 +20,3 @@ class FavaAPIException(Exception):
 
     def __str__(self) -> str:
         return self.message
-
-
-class FavaModule:
-    """Base class for the "modules" of FavaLedger."""
-
-    def __init__(self, ledger):
-        self.ledger = ledger
-
-    def load_file(self):
-        """Gets called when the file has been (re)loaded."""
